+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file, formatted via [this recommendation](http://keepachangelog.com/).
 
+## [1.5.0] - 2015-11-XX
+### Added
+- Email sharing feature
+- Plugin updater
+
 ## [1.4.0] - 2015-10-21
 ### Added
 - Print included service option